refactor(ui): split code into multiple files
