+++ shell/crates/running_apps/src/ui/components/mod.rs
@@ -1,0 +1,3 @@
+pub mod footer;
+pub mod card;
+pub mod scroller;
